Refactor controls.
